+++ operators/advanced_network/advanced_network/managers/gpunetio/adv_network_doca_mgr.h
@@ -217,8 +217,7 @@ class DocaMgr : public Manager {
   void free_all_packets(BurstParams* burst) override{};
   void free_rx_burst(BurstParams* burst) override;
   void free_tx_burst(BurstParams* burst) override;
-  std::optional<uint16_t> get_port_from_ifname(const std::string& name) override;
-
+  
   Status get_rx_burst(BurstParams** burst) override;
   Status set_packet_tx_time(BurstParams* burst, int idx, uint64_t timestamp);
   void free_rx_metadata(BurstParams* burst) override;